refactor: split handlers monolith into a handlers/ package
Break the 1100-line handlers.py into a package:
- handlers/_helpers.py: pure helpers, constants, TypedDicts, and the
  @_rule_handler registry primitives.
- handlers/registry.py: the HandlerRegistry dispatch class and
  generic_handler entry point.
- handlers/__init__.py: preserves the public import surface.

Behavior is unchanged. Test mock-patch targets for names now looked up in
the registry submodule namespace (resolve_target_value,
_read_project_python_file) are updated accordingly.

Refs #190
