+++ tests/test_handler.py
@@ -83,7 +83,9 @@ def test_compare_func_core_tools_version_pass(monkeypatch: MonkeyPatch) -> None:
     from azure_functions_doctor import handlers
 
     monkeypatch.setattr(
-        handlers, "resolve_target_value", lambda t: "4.0.5455" if t == "func_core_tools" else ""
+        handlers.registry,
+        "resolve_target_value",
+        lambda t: "4.0.5455" if t == "func_core_tools" else "",
     )
     rule: Rule = {
         "type": "compare_version",
@@ -103,7 +105,7 @@ def test_compare_func_core_tools_version_fail_not_installed(monkeypatch: MonkeyP
     from azure_functions_doctor import handlers
 
     monkeypatch.setattr(
-        handlers,
+        handlers.registry,
         "resolve_target_value",
         lambda t: "not_installed" if t == "func_core_tools" else "",
     )
@@ -125,7 +127,9 @@ def test_compare_func_core_tools_version_fail_old_version(monkeypatch: MonkeyPat
     from azure_functions_doctor import handlers
 
     monkeypatch.setattr(
-        handlers, "resolve_target_value", lambda t: "3.0.0" if t == "func_core_tools" else ""
+        handlers.registry,
+        "resolve_target_value",
+        lambda t: "3.0.0" if t == "func_core_tools" else "",
     )
     rule: Rule = {
         "type": "compare_version",
