+++ tests/test_handler_registry_extended.py
@@ -51,7 +51,7 @@ def test_compare_version_invalid_version() -> None:
     }
     # Mock func_core_tools to return invalid version
     # Mock func_core_tools to return invalid version
-    _target = "azure_functions_doctor.handlers.resolve_target_value"
+    _target = "azure_functions_doctor.handlers.registry.resolve_target_value"
     with patch(_target, return_value="invalid!!!version"):
         result = registry.handle(rule, Path("."))
         assert result["status"] == "fail"
@@ -324,7 +324,7 @@ def test_conditional_exists_durable_detection_exception() -> None:
             raise ValueError("Simulated read error")
 
         with patch(
-            "azure_functions_doctor.handlers._read_project_python_file",
+            "azure_functions_doctor.handlers.registry._read_project_python_file",
             side_effect=read_with_error,
         ):
             result = registry.handle(rule, tmp_path)
